ml/requirements.txt
requests
scikit-learn
numpy
pandas
matplotlib
seaborn
textblob
flask
flask-cors
python-dotenv
scipy
gunicorn
